refactor: reduce packages (remove all @terra-dev/*)
