+++ packages/.packages.json
@@ -1,9 +1,5 @@
 {
   "$schema": "https://rocket-hangar.github.io/rocket-punch/schemas/packages.json",
-  "@terra-dev/*": {
-    "version": "3.0.0-alpha.1",
-    "tag": "next"
-  },
   "@terra-money/*": {
     "version": "3.0.0-alpha.1",
     "tag": "next"
