--- packages/src/@terra-dev/chrome-extension/index.ts
@@ -1,4 +1,0 @@
-export * from './ChromeExtensionController';
-export * from './types';
-export * from './errors';
-export * from './multiChannel';
