--- packages/src/@terra-dev/readonly-wallet-modal/README.md
@@ -1,3 +1,0 @@
-# `@terra-dev/readonly-wallet-modal`
-
-This component is the default modal component of [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider). You don't have to install it.
