--- packages/src/@terra-dev/readonly-wallet/README.md
@@ -1,10 +1,0 @@
-# Deprecation warning
-
-> We think this is an unnecessarily separated package. So this package will be removed in the next major release. please use `@terra-money/wallet-provider/modules/readonly-wallet`.
-
-# `@terra-dev/readonly-wallet`
-
-This package is one of the components
-of [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider). If there is no
-particular reason, please
-use [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider).
