--- packages/src/@terra-dev/readonly-wallet/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * from './connect';
