--- packages/src/@terra-dev/wallet-types/README.md
@@ -1,9 +1,0 @@
-# `@terra-dev/wallet-types`
-
-This is a package commonly referenced
-by [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider) and its subconfiguration
-packages. You don't need to use it if there's no other reason.
-
-All the items in this package are being re-exported from
-the [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider), so you don't have to
-install it separately.
