--- packages/src/@terra-dev/wallet-types/index.ts
@@ -1,2 +1,0 @@
-export * from './errors';
-export * from './types';
