--- packages/src/@terra-dev/walletconnect-qrcode-modal/README.md
@@ -1,3 +1,0 @@
-# `@terra-dev/walletconnect-qrcode-modal`
-
-This component is the default modal component of [`@terra-money/wallet-provider`](https://www.npmjs.com/package/@terra-money/wallet-provider). You don't have to install it.
