--- packages/src/@terra-dev/walletconnect/index.ts
@@ -1,4 +1,0 @@
-export * from './connect';
-export * from './types';
-export * from './errors';
-export * from './impl/socket-transport';
