+++ packages/src/@terra-money/use-wallet/index.ts
@@ -1,4 +1,5 @@
 export * from './tx';
 export * from './types';
+export * from './errors';
 export * from './useWallet';
 export * from './useConnectedWallet';
