+++ packages/src/@terra-money/use-wallet/useConnectedWallet.ts
@@ -1,8 +1,12 @@
-import { NetworkInfo, SignResult } from '@terra-dev/wallet-types';
 import { AccAddress, CreateTxOptions } from '@terra-money/terra.js';
 import { useMemo } from 'react';
-import { TxResult } from './tx';
-import { ConnectType, WalletStatus } from './types';
+import {
+  ConnectType,
+  NetworkInfo,
+  SignResult,
+  TxResult,
+  WalletStatus,
+} from './types';
 import { useWallet } from './useWallet';
 
 type HumanAddr = string & { __type: 'HumanAddr' };
