+++ packages/src/@terra-money/wallet-provider/index.ts
@@ -2,5 +2,5 @@ export * from './getChainOptions';
 export * from './controller';
 export * from './react';
 
-export * from '@terra-dev/use-wallet';
-export * from '@terra-dev/wallet-types';
+export * from '@terra-money/use-wallet';
+export type { ReadonlyWalletSession } from './modules/readonly-wallet';
