+++ packages/src/@terra-money/wallet-provider/modules/chrome-extension/errors.ts
