+++ packages/src/@terra-money/wallet-provider/modules/chrome-extension/index.ts
@@ -1,17 +1,4 @@
-import {
-  ChromeExtensionController,
-  ChromeExtensionControllerOptions,
-  ChromeExtensionCreateTxFailed,
-  ChromeExtensionStatus,
-  ChromeExtensionTxFailed,
-  ChromeExtensionUnspecifiedError,
-} from '@terra-dev/chrome-extension';
-
-export {
-  ChromeExtensionController,
-  ChromeExtensionStatus,
-  ChromeExtensionCreateTxFailed,
-  ChromeExtensionTxFailed,
-  ChromeExtensionUnspecifiedError,
-};
-export type { ChromeExtensionControllerOptions };
+export * from './ChromeExtensionController';
+export * from './types';
+export * from './errors';
+export * from './multiChannel';
