+++ packages/src/@terra-money/wallet-provider/modules/readonly-wallet/index.ts
@@ -1,14 +1,3 @@
-import {
-  connect,
-  connectIfSessionExists,
-  ReadonlyWalletController,
-  ReadonlyWalletOptions,
-  ReadonlyWalletSession,
-} from '@terra-dev/readonly-wallet';
-
-export { connectIfSessionExists, connect };
-export type {
-  ReadonlyWalletOptions,
-  ReadonlyWalletController,
-  ReadonlyWalletSession,
-};
+export * from './types';
+export * from './connect';
+export * from './modal';
