+++ packages/src/@terra-money/wallet-provider/modules/walletconnect/errors.ts
