+++ packages/src/@terra-money/wallet-provider/modules/walletconnect/impl/socket-transport/network.ts
