+++ packages/src/@terra-money/wallet-provider/react/useInstallChromeExtension.ts
@@ -1,4 +1,4 @@
-import { ConnectType, useWallet } from '@terra-dev/use-wallet';
+import { ConnectType, useWallet } from '@terra-money/use-wallet';
 import { useMemo } from 'react';
 
 export function useInstallChromeExtension() {
